Added more info on creating combo charts
I have bolded the the instruction on creating combo charts.
I have added an additional warning on creating a combo chart. I am not sure how it will look on the site as the Previewer does not render correctly. But it should be at least a note. Also I am not sure how the lone sentence between the to notes will look.
I have added a small code snippet.