Update create-custom-image-properties-resolver-net-standard.md
Much of the TryGetRawImageData method was not working due to it trying to use features of the SixLabors.ImageSharp which are not available in the current version, mostly it looks like they changed the decoders into static classes/methods so the new keyword is not working there. I've dropped in a version that works with SixLabors.ImageSharp version 3.1.4.